# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFG API Segura - Gestión de Informes Médicos",
        "description": "API REST segura con autenticación JWT, MFA, roles y permisos, rate limiting y auditoría de seguridad.",
        "contact": {
            "name": "Pablo Sánchez",
            "email": "pablo@tfg.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/informes": {
            "get": {
                "tags": [
                    "Informes"
                ],
                "summary": "Listar informes médicos",
                "operationId": "96300eb9610cc29b15052a3ddb7ca8e8",
                "parameters": [
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Filtrar por título",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paciente_id",
                        "in": "query",
                        "description": "Filtrar por paciente (solo admin y manager)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de informes paginada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Informes"
                ],
                "summary": "Crear un nuevo informe médico",
                "operationId": "980b66394c0cd7136e8a1af9caa169c2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "diagnostico",
                                    "paciente_id",
                                    "medico_id"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Informe de revisión"
                                    },
                                    "diagnostico": {
                                        "type": "string",
                                        "example": "Paciente en buen estado"
                                    },
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "archivo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Informe creado con éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Datos incorrectos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/informes/{id}": {
            "get": {
                "tags": [
                    "Informes"
                ],
                "summary": "Ver un informe médico",
                "operationId": "a2482eebf1df830d5860166d4c9a3bbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informe encontrado"
                    },
                    "403": {
                        "description": "Acceso no autorizado"
                    },
                    "404": {
                        "description": "Informe no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Informes"
                ],
                "summary": "Actualizar un informe médico",
                "operationId": "1e006149672068d9baddfd029862b3cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Informe actualizado"
                                    },
                                    "diagnostico": {
                                        "type": "string",
                                        "example": "Nuevo diagnóstico"
                                    },
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Informe actualizado"
                    },
                    "403": {
                        "description": "Acceso no autorizado"
                    },
                    "404": {
                        "description": "Informe no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Informes"
                ],
                "summary": "Eliminar un informe médico",
                "operationId": "4250ada9b078261c3b0ade1847afcd72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informe eliminado"
                    },
                    "403": {
                        "description": "Solo administradores pueden eliminar"
                    },
                    "404": {
                        "description": "Informe no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}